% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveAllocation.R
\docType{package}
\name{AdAlloc}
\alias{AdaptiveAllocation}
\alias{AdaptiveAllocation-package}
\alias{AdAlloc}
\title{Adaptive Allocation}
\description{
The implementation has been done in \code{C++} through the use of
\code{Rcpp} and \code{RcppArmadillo}.
}
\references{
TBA
}
\author{
Davide Fabbrico

Maintainer: Davide Fabbrico \email{davide.fabbrico@unifi.it}
}
