% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{AdaptCRSG}
\alias{AdaptCRSG}
\title{Categorical Diversity-Guided Scan Gibbs}
\usage{
AdaptCRSG(
  X,
  hyper = c(1, 1, 1),
  K = 3,
  R = 3,
  m = 10,
  iteration = 1000,
  burnin = 50,
  iterTuning = 50,
  thin = 5,
  updateProbAllocation = 1,
  method = "",
  q = 1,
  lambda = 1,
  kWeibull = 1,
  alphaPareto = 1,
  xmPareto = 0.5,
  DiversityIndex = "Exponential",
  adaptive = FALSE,
  nSD = 1,
  lambda0 = 40,
  zeta = 0.996,
  a = 100
)
}
\description{
Categorical Diversity-Guided Scan Gibbs
}
