% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{AdaptRSG}
\alias{AdaptRSG}
\title{Diversity-Guided Gibbs Sampler}
\usage{
AdaptRSG(
  X,
  hyper = c(1, 1, 0, 1, 1, 1),
  K = 3,
  m = 10,
  iteration = 1000,
  burnin = 50,
  thin = 5,
  updateProbAllocation = 1,
  method = "",
  q = 1,
  lambda = 1,
  kWeibull = 1,
  alphaPareto = 1,
  xmPareto = 0.5,
  DiversityIndex = "Exponential",
  adaptive = FALSE,
  nSD = 1.96,
  lambda0 = 30,
  L = 1,
  max_lambda = 50,
  c = 1,
  a = 1,
  w_fun = "hyperbolic",
  sp = 0,
  seed = 10
)
}
\description{
Diversity-Guided Gibbs Sampler
}
