% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{cdsg}
\alias{cdsg}
\title{Categorical Diversity-Guided Scan Gibbs}
\usage{
cdsg(
  X,
  hyper = c(1, 1, 1),
  K = 3,
  R = 3,
  m = 10,
  iteration = 1000,
  burnin = 50,
  thin = 5,
  gamma = 0.5,
  trueAllocation = numeric(0),
  adaptiveGamma = FALSE,
  q = 1
)
}
\description{
Categorical Diversity-Guided Scan Gibbs
}
