% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{crsg}
\alias{crsg}
\title{Categorical Random Scan Gibbs}
\usage{
crsg(
  X,
  hyper = c(1, 1, 1),
  K = 3,
  R = 3,
  m = 10,
  iteration = 1000,
  burnin = 50,
  thin = 5
)
}
\description{
Categorical Random Scan Gibbs
}
