% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{cssg}
\alias{cssg}
\title{Categorical Systematic Scan Gibbs}
\usage{
cssg(
  X,
  hyper = c(1, 1, 1),
  K = 3,
  R = 3,
  iteration = 1000,
  burnin = 50,
  thin = 5,
  trueAllocation = numeric(0)
)
}
\description{
Categorical Systematic Scan Gibbs
}
