% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{EntropyGibbs}
\alias{EntropyGibbs}
\title{Entropy-Guided Gibbs Sampler}
\usage{
EntropyGibbs(
  X,
  hyper = c(1, 1, 0, 20, 1, 1),
  K = 3,
  m,
  iteration = 1000,
  burnin = 50,
  thin = 5,
  method = "EB",
  gamma = 0.5
)
}
\description{
Entropy-Guided Gibbs Sampler
}
