% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psm}
\alias{psm}
\title{Posterior Summary - Bayesian nonparametric mixture inconsistency for the number of components: 
How worried should we be in practice? - by Paul Kirk page 7.}
\usage{
psm(z)
}
\description{
Posterior Summary - Bayesian nonparametric mixture inconsistency for the number of components: 
How worried should we be in practice? - by Paul Kirk page 7.
}
