% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{rssg}
\alias{rssg}
\title{rssg}
\usage{
rssg(
  X,
  hyper = c(1, 1, 0, 1, 1, 1),
  K = 3,
  m = 10,
  iteration = 1000,
  burnin = 50,
  thin = 5,
  method = "",
  trueAllocation = numeric(0)
)
}
\description{
rssg
}
