% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{AdaptRSG}
\alias{AdaptRSG}
\title{Diversity-Guided Gibbs Sampler}
\usage{
AdaptRSG(
  X,
  hyper = c(1, 1, 0, 1, 1, 1),
  K = 3,
  m = 10,
  iteration = 1000,
  burnin = 50,
  iterTuning = 50,
  thin = 5,
  updateProbAllocation = 1,
  method = "",
  gamma = 0.5,
  q = 1,
  lambda = 1,
  kWeibull = 1,
  alphaPareto = 1,
  xmPareto = 0.5,
  DiversityIndex = "Exponential",
  adaptive = FALSE,
  nSD = 1,
  lambda0 = 40,
  zeta = 0.996,
  a = 100
)
}
\description{
Diversity-Guided Gibbs Sampler
}
