% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCpp.R
\name{DiversityGibbs}
\alias{DiversityGibbs}
\title{Diversity-Guided Gibbs Sampler}
\usage{
DiversityGibbs(
  X,
  hyper = c(1, 1, 0, 0.1, 1, 1),
  K = 3,
  m = 10,
  iteration = 1000,
  burnin = 50,
  thin = 5,
  method = "EB",
  gamma = 0.5,
  diversity = "Entropy",
  trueAllocation = numeric(0)
)
}
\description{
Diversity-Guided Gibbs Sampler
}
